Handle calls when Netp is enabled only
